Rearranging some code
